Introduction doi of the publication, some text edit
With the addition of DataWarrior's web site, this work is set
into the context as a plugin of an already existing software.  The
addition of the publication's doi offers an easier access to the
relevant publication, and follow-up how this sets into context
as a reference about tautomers.

The consecution of the paragraphs was altered into a pattern of
"origin / background of this project" -> "what does the project to the
table" -> "outlook / further development of the project".

A line break in an text editor does not automatically equate an
empty line in github's presentation of the markdown file.  However,
it offers an easier track of changes in a difference view, e.g.
by vimdiff, or in a simultaneous display of multiple versions of the
file in a GUI side-by-side.